+++ instrumentation/othttp/handler_example_test.go
@@ -24,12 +24,9 @@ import (
 
 	"go.opentelemetry.io/otel/api/kv"
 
-	"go.opentelemetry.io/otel/api/global"
 	"go.opentelemetry.io/otel/api/trace"
-	mstdout "go.opentelemetry.io/otel/exporters/metric/stdout"
-	"go.opentelemetry.io/otel/exporters/trace/stdout"
+	"go.opentelemetry.io/otel/exporters/stdout"
 	"go.opentelemetry.io/otel/instrumentation/othttp"
-	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 )
 
 func ExampleNewHandler() {
@@ -48,29 +45,14 @@ func ExampleNewHandler() {
 	*/
 
 	// Write spans to stdout
-	exporter, err := stdout.NewExporter(stdout.Options{PrettyPrint: true})
+	pusher, err := stdout.InstallNewPipeline([]stdout.Option{
+		stdout.WithPrettyPrint(),
+		stdout.WithoutTimestamps(), // This makes the output deterministic
+	}, nil)
 	if err != nil {
 		log.Fatal(err)
 	}
-
-	tp, err := sdktrace.NewProvider(sdktrace.WithConfig(sdktrace.Config{DefaultSampler: sdktrace.AlwaysSample()}),
-		sdktrace.WithSyncer(exporter))
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	pusher, err := mstdout.NewExportPipeline(mstdout.Config{
-		PrettyPrint:    true,
-		DoNotPrintTime: true, // This makes the output deterministic
-	})
-
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	meterProvider := pusher.Provider()
-	global.SetTraceProvider(tp)
-	global.SetMeterProvider(meterProvider)
+	defer pusher.Stop()
 
 	figureOutName := func(ctx context.Context, s string) (string, error) {
 		pp := strings.SplitN(s, "/", 2)